[vuex] add auth module
