+++ src/components/SpSignIn.vue
@@ -210,7 +210,7 @@ export default {
 			return bip39.validateMnemonic(this.mnemonicClean)
 		},
 		address() {
-			const { client } = this.$store.state.cosmos.bank
+			const client = this.$store.getters['cosmos/client']
 			return client && client.senderAddress
 		}
 	},
